+++ packages/dnb-ui-lib/src/components/lib.js
@@ -44,6 +44,7 @@ import Switch from './switch/Switch'
 import Tabs from './tabs/Tabs'
 import Textarea from './textarea/Textarea'
 import ToggleButton from './toggle-button/ToggleButton'
+import Tooltip from './tooltip/Tooltip'
 
 // define / export all the aviable components
 export {
@@ -78,7 +79,8 @@ export {
   Switch,
   Tabs,
   Textarea,
-  ToggleButton
+  ToggleButton,
+  Tooltip
 }
 
 export const getComponents = () => {
@@ -114,7 +116,8 @@ export const getComponents = () => {
     Switch,
     Tabs,
     Textarea,
-    ToggleButton
+    ToggleButton,
+    Tooltip
   }
 }