+++ packages/dnb-ui-lib/src/index.js
@@ -42,6 +42,7 @@ import Switch from './components/switch/Switch'
 import Tabs from './components/tabs/Tabs'
 import Textarea from './components/textarea/Textarea'
 import ToggleButton from './components/toggle-button/ToggleButton'
+import Tooltip from './components/tooltip/Tooltip'
 import Anchor from './elements/Anchor'
 import Blockquote from './elements/Blockquote'
 import Code from './elements/Code'
@@ -98,6 +99,7 @@ export {
   Tabs,
   Textarea,
   ToggleButton,
+  Tooltip,
   Anchor,
   Blockquote,
   Code,