+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - The `DNB` font got updates on the bold weight, mostly effected Windows users
 - New component [Autocomplete](/uilib/components/autocomplete)
 - New component [Pagination](/uilib/components/pagination) including infinity scroller
+- New component [Heading](/uilib/components/heading)
 - New [secondary icons](/icons/secondary): `pay_from`, `transfer_to`
 - Modal got a [Drawer mode](/uilib/components/modal/demos#drawer-mode) and has now a dark background color
 - **Breaking Changes** to the UMD [bundles](/uilib/usage/first-steps/bundles) structure (v7), including ESM (mjs)